Pitchain/backend/requirements.txt
Django==4.2.10
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
python-decouple==3.8
requests==2.31.0
web3==6.15.1
Pillow==11.1.0
django-filter==23.5
